i386: Revised handling for unknown frame types.
The existing scheme was biassed heavily to the assumption that unknown frame
types would be "normal" (i.e. that %ebp would point to valid data).  This fails
in some very common cases on Darwin - when we are skipping past calls to libc.

We step into a stub handler - which is essentially an indirect jump to the
dynamic linker lazy symbol resolution mechanism.  These stubs have no %ebp which
then means that the value returned points not to the caller's stack frame but to
the caller's caller's (i.e. effectively we drop on stack frame from the list).

This then means that we do not get the correct caller's frame PC and we cannot
either step past calls to different DSOs sucessfully past or print a backtrace.

I am kinda surprised this does not affect other targets in a similar way.

The update here modifies the handling to match that used for x86-64, where we do
not make any such assumptions about $rbp validity for unknown stack frame types.

TODO, maybe: we could invent a new unwind sniffer that considers stub code as
something different (but adding a new FRAME type looks like quite heavy lifting).